refactor: variable names in bulk action subscription app
