+++ admin/src/shared/components/BulkAction/entities/subscriptionMenuProduct.js
@@ -1,25 +1,26 @@
 import React from 'react'
 import {
+   Spacer,
    TextButton,
    Text,
    Flex,
    Form,
-   HorizontalTabs,
-   HorizontalTabList,
-   HorizontalTabPanels,
-   HorizontalTab,
    HorizontalTabPanel,
+   HorizontalTabPanels,
+   HorizontalTabList,
+   HorizontalTab,
+   HorizontalTabs,
    Collapsible,
-   Spacer,
+   Dropdown,
    ButtonGroup,
    HelperText,
-   Dropdown,
 } from '@dailykit/ui'
 import { Tooltip } from '../..'
-import { PRODUCT_CATEGORY_OF_SUBSCRIPTION_OCCURRENCE_PRODUCT } from '../subscription'
 import { useSubscription } from '@apollo/react-hooks'
 
-export const ManageAddToSubscriptionMenu = props => {
+import { PRODUCT_CATEGORY_OF_SUBSCRIPTION_OCCURRENCE_PRODUCT } from '../subscription'
+
+export const SubscriptionOccurrenceProductBulkAction = props => {
    const {
       initialBulkAction,
       setInitialBulkAction,
@@ -30,6 +31,7 @@ export const ManageAddToSubscriptionMenu = props => {
    } = props
    const [productCategories, setProductCategories] = React.useState([])
 
+   // subscription
    useSubscription(PRODUCT_CATEGORY_OF_SUBSCRIPTION_OCCURRENCE_PRODUCT, {
       onSubscriptionData: data => {
          console.log(
@@ -103,7 +105,7 @@ export const ManageAddToSubscriptionMenu = props => {
             placeholder="choose product type"
          />
          <Spacer size="20px" />
-         {[{ heading: 'Unit Price', columnName: 'unitPrice' }].map(
+         {[{ heading: 'Add On Price', columnName: 'addOnPrice' }].map(
             (column, i) => (
                <CollapsibleComponent heading={column.heading} key={i}>
                   <HorizontalTabs>
@@ -119,7 +121,7 @@ export const ManageAddToSubscriptionMenu = props => {
                                  console.log('clear')
                               }}
                            >
-                              Set {column.columnName}
+                              Set {column.heading}
                            </TextButton>
                         </HorizontalTab>
                         <HorizontalTab>
@@ -139,8 +141,8 @@ export const ManageAddToSubscriptionMenu = props => {
                         <HorizontalTabPanel>
                            <Form.Group>
                               <Form.Label
-                                 htmlFor="UnitPrice "
-                                 title="Unit Price "
+                                 htmlFor="AddOnPrice "
+                                 title="Add On Price "
                               >
                                  <Flex container alignItems="center">
                                     <Text as="text1">{column.heading}</Text>
@@ -178,6 +180,7 @@ export const ManageAddToSubscriptionMenu = props => {
                                  id={column.columnName}
                                  name={column.columnName}
                                  min="0"
+                                 // disabled={initialBulkAction.addOnPrice.decrease !== 0}
                                  value={
                                     initialBulkAction[column.columnName].set
                                  }
@@ -408,7 +411,17 @@ export const ManageAddToSubscriptionMenu = props => {
                </CollapsibleComponent>
             )
          )}
-
+         <CollapsibleComponentWithTabs
+            bulkActions={bulkActions}
+            columnName="addOnLabel"
+            title="Add On Label"
+            concatType="concatDataString"
+            columnConcat="addOnLabelConcat"
+            initialBulkAction={initialBulkAction}
+            setBulkActions={setBulkActions}
+            setInitialBulkAction={setInitialBulkAction}
+            isNullable={false}
+         />
          <Spacer size="20px" />
 
          <Form.Group>
@@ -479,6 +492,439 @@ export const ManageAddToSubscriptionMenu = props => {
             </Form.Toggle>
          </Form.Group>
       </>
+   )
+}
+const CollapsibleComponentWithTabs = ({
+   bulkActions,
+   columnName,
+   title,
+   concatType,
+   columnConcat,
+   example,
+   initialBulkAction,
+   isNullable,
+   setBulkActions,
+   setInitialBulkAction,
+}) => {
+   const capitalize = string => string.charAt(0).toUpperCase() + string.slice(1)
+   const checkNested = (obj, level, ...rest) => {
+      if (obj === undefined) return false
+      if (rest.length == 0 && obj.hasOwnProperty(level)) return true
+      return checkNested(obj[level], ...rest)
+   }
+   const commonRemove = (
+      column,
+      concatTypeCR,
+      positionPrimary,
+      positionSecondary
+   ) => {
+      //positionPrimary use for changing field
+      //positionSecondary use for another field
+      //concatTypeCR -> concatType in common remove function
+      concatTypeCR =
+         concatTypeCR === 'concatData' ? 'concatData' : 'concatDataString'
+
+      const nestedCheckPrepend = checkNested(
+         bulkActions,
+         concatTypeCR,
+         column,
+         positionSecondary
+      )
+      if (nestedCheckPrepend) {
+         const newBulkAction = { ...bulkActions }
+         delete newBulkAction[concatTypeCR][column][positionPrimary]
+         setBulkActions(newBulkAction)
+         return
+      }
+      const checkNestedColumn = checkNested(bulkActions, concatTypeCR, column)
+      if (checkNestedColumn) {
+         const newBulkAction = { ...bulkActions }
+         delete newBulkAction[concatTypeCR][column]
+         setBulkActions(newBulkAction)
+      }
+      if (concatTypeCR in bulkActions) {
+         if (Object.keys(bulkActions[concatTypeCR]).length === 0) {
+            const newBulkAction = { ...bulkActions }
+            delete newBulkAction[concatTypeCR]
+            setBulkActions(newBulkAction)
+         }
+      }
+   }
+   const isSetVisible = (column, concatTypeISV) => {
+      //concatTypeCR -> concatType in is visible function
+
+      concatTypeISV =
+         concatTypeISV === 'concatData' ? 'concatData' : 'concatDataString'
+
+      const checkForColumnNull = initialBulkAction[column] == null
+      if (checkForColumnNull) return checkForColumnNull
+      const checkForAppendPrepend =
+         checkNested(bulkActions, concatTypeISV, column, 'appendvalue') ||
+         checkNested(bulkActions, concatTypeISV, column, 'prependvalue')
+      return checkForAppendPrepend || checkForColumnNull
+   }
+   const isAppendPrependVisible = column => {
+      const checkForColumnNull = initialBulkAction[column] == null
+      if (checkForColumnNull) return checkForColumnNull
+      const checkForColumnLength = initialBulkAction[column].length > 0
+      return checkForColumnNull || checkForColumnLength
+   }
+   const isSetNullVisible = (column, concatTypeISV) => {
+      //concatTypeCR -> concatType in is visible function
+      concatTypeISV =
+         concatTypeISV === 'concatData' ? 'concatData' : 'concatDataString'
+      if (!(initialBulkAction[column] == null)) {
+         const checkForAppendPrepend =
+            checkNested(bulkActions, concatTypeISV, column, 'appendvalue') ||
+            checkNested(bulkActions, concatTypeISV, column, 'prependvalue')
+         const checkForColumnLength = initialBulkAction[column].length > 0
+         return checkForAppendPrepend || checkForColumnLength
+      }
+   }
+   return (
+      <CollapsibleComponent heading={title}>
+         <HorizontalTabs>
+            <HorizontalTabList>
+               <HorizontalTab>
+                  <TextButton
+                     type="ghost"
+                     size="sm"
+                     disabled={isSetVisible(columnName, concatType)}
+                     onClick={() => {}}
+                  >
+                     Set {title}
+                  </TextButton>
+               </HorizontalTab>
+               <HorizontalTab>
+                  <TextButton
+                     type="ghost"
+                     size="sm"
+                     disabled={isAppendPrependVisible(columnName, concatType)}
+                     onClick={() => {}}
+                  >
+                     Append or Prepend
+                  </TextButton>
+               </HorizontalTab>
+               {isNullable && (
+                  <HorizontalTab>
+                     <TextButton
+                        type="ghost"
+                        size="sm"
+                        disabled={isSetNullVisible(columnName, concatType)}
+                        onClick={() => {}}
+                     >
+                        Set Null
+                     </TextButton>
+                  </HorizontalTab>
+               )}
+            </HorizontalTabList>
+            <HorizontalTabPanels>
+               <HorizontalTabPanel>
+                  <Form.Group>
+                     <Form.Label htmlFor={columnName} title={columnName}>
+                        <Flex container alignItems="center">
+                           <Text as="text1">{title}</Text>
+                           <TextButton
+                              type="ghost"
+                              size="sm"
+                              onClick={() => {
+                                 setInitialBulkAction({
+                                    ...initialBulkAction,
+                                    [columnName]: '',
+                                 })
+                                 setBulkActions(prevState => {
+                                    const newOption = { ...prevState }
+                                    delete newOption[columnName]
+                                    return newOption
+                                 })
+                              }}
+                           >
+                              Clear
+                           </TextButton>
+                           <Tooltip identifier="products" />
+                        </Flex>
+                     </Form.Label>
+                     <Form.Text
+                        id={columnName}
+                        name={columnName}
+                        value={initialBulkAction[columnName]}
+                        onChange={e =>
+                           setInitialBulkAction({
+                              ...initialBulkAction,
+                              [columnName]: e.target.value,
+                           })
+                        }
+                        onBlur={() => {
+                           if (initialBulkAction[columnName]) {
+                              const newColumnName =
+                                 initialBulkAction[columnName]
+                              setBulkActions({
+                                 ...bulkActions,
+                                 [columnName]: newColumnName,
+                              })
+                              return
+                           }
+                           if (columnName in bulkActions) {
+                              const newOptions = { ...bulkActions }
+                              delete newOptions[columnName]
+                              setBulkActions(newOptions)
+                              return
+                           }
+                        }}
+                        placeholder={`Enter ${title}`}
+                     />
+                  </Form.Group>
+                  <Form.Error>
+                     Changing Add on Label will overwrite already existing Add
+                     on Label
+                  </Form.Error>
+                  {concatType == 'concatData' && (
+                     <HelperText
+                        message={`Enter comma separated values, for example: ${example}`}
+                        type="hint"
+                     />
+                  )}
+               </HorizontalTabPanel>
+               <HorizontalTabPanel>
+                  <Form.Group>
+                     <Form.Label htmlFor={columnName} title={columnName}>
+                        <Flex container alignItems="center">
+                           <Text as="text1">{title} Append</Text>
+                           <TextButton
+                              type="ghost"
+                              size="sm"
+                              onClick={() => {
+                                 setInitialBulkAction({
+                                    ...initialBulkAction,
+                                    [columnConcat]: {
+                                       ...initialBulkAction[columnConcat],
+                                       forAppend: '',
+                                    },
+                                 })
+                                 commonRemove(
+                                    columnName,
+                                    concatType,
+                                    'appendvalue',
+                                    'prependvalue'
+                                 )
+                              }}
+                           >
+                              Clear
+                           </TextButton>
+                           <Tooltip identifier="product" />
+                        </Flex>
+                     </Form.Label>
+                     <Form.Text
+                        id={`${columnName}Append`}
+                        name={`${columnName}Append`}
+                        value={initialBulkAction[columnConcat].forAppend}
+                        onChange={e =>
+                           setInitialBulkAction({
+                              ...initialBulkAction,
+                              [columnConcat]: {
+                                 ...initialBulkAction[columnConcat],
+                                 forAppend: e.target.value,
+                              },
+                           })
+                        }
+                        onBlur={() => {
+                           if (initialBulkAction[columnConcat].forAppend) {
+                              let newColumnAppend
+                              if (concatType == 'concatData') {
+                                 newColumnAppend = initialBulkAction[
+                                    columnConcat
+                                 ].forAppend
+                                    .split(',')
+                                    .map(tag => {
+                                       const newTag = tag.trim()
+                                       return capitalize(newTag)
+                                    })
+                              } else {
+                                 newColumnAppend =
+                                    initialBulkAction[columnConcat].forAppend
+                              }
+
+                              const newConcatData = {
+                                 ...bulkActions[concatType],
+                              }
+                              const newColumnName = {
+                                 ...newConcatData[columnName],
+                              }
+                              newColumnName.columnname = columnName
+                              newColumnName.appendvalue = newColumnAppend
+                              newColumnName.schemaname = 'subscription'
+                              newColumnName.tablename =
+                                 'subscriptionOccurence_product'
+                              newConcatData[columnName] = newColumnName
+                              setBulkActions({
+                                 ...bulkActions,
+                                 [concatType]: newConcatData,
+                              })
+                              return
+                           }
+                           commonRemove(
+                              columnName,
+                              columnName,
+                              'appendvalue',
+                              'prependvalue'
+                           )
+                        }}
+                        placeholder={`Enter append ${columnName}`}
+                     />
+                  </Form.Group>
+                  {concatType == 'concatData' && (
+                     <HelperText
+                        type="hint"
+                        message={`Enter comma separated values, for example: ${example}`}
+                     />
+                  )}
+                  <Form.Group>
+                     <Form.Label htmlFor={columnName} title={columnName}>
+                        <Flex container alignItems="center">
+                           <Text as="text1">{title} Prepend</Text>
+                           <TextButton
+                              type="ghost"
+                              size="sm"
+                              onClick={() => {
+                                 setInitialBulkAction({
+                                    ...initialBulkAction,
+                                    [columnConcat]: {
+                                       ...initialBulkAction[columnConcat],
+                                       forPrepend: '',
+                                    },
+                                 })
+                                 commonRemove(
+                                    columnName,
+                                    concatType,
+                                    'prependvalue',
+                                    'appendvalue'
+                                 )
+                              }}
+                           >
+                              Clear
+                           </TextButton>
+                           <Tooltip identifier="products" />
+                        </Flex>
+                     </Form.Label>
+                     <Form.Text
+                        id={`${columnName}Prepend`}
+                        name={`${columnName}Prepend`}
+                        value={initialBulkAction[columnConcat].forPrepend}
+                        onChange={e =>
+                           setInitialBulkAction({
+                              ...initialBulkAction,
+                              [columnConcat]: {
+                                 ...initialBulkAction[columnConcat],
+                                 forPrepend: e.target.value,
+                              },
+                           })
+                        }
+                        onBlur={() => {
+                           if (initialBulkAction[columnConcat].forPrepend) {
+                              let newColumnPrepend
+                              if (concatType == 'concatData') {
+                                 newColumnPrepend = initialBulkAction[
+                                    columnConcat
+                                 ].forPrepend
+                                    .split(',')
+                                    .map(tag => {
+                                       const newTag = tag.trim()
+                                       return capitalize(newTag)
+                                    })
+                              } else {
+                                 newColumnPrepend =
+                                    initialBulkAction[columnConcat].forPrepend
+                              }
+
+                              const newConcatData = {
+                                 ...bulkActions[concatType],
+                              }
+                              const newColumnName = {
+                                 ...newConcatData[columnName],
+                              }
+                              newColumnName.columnname = columnName
+                              newColumnName.prependvalue = newColumnPrepend
+                              newColumnName.schemaname = 'subscription'
+                              newColumnName.tablename =
+                                 'subscriptionOccurence_product'
+                              newConcatData[columnName] = newColumnName
+                              setBulkActions({
+                                 ...bulkActions,
+                                 [concatType]: newConcatData,
+                              })
+                              return
+                           }
+                           commonRemove(
+                              columnName,
+                              concatType,
+                              'prependvalue',
+                              'appendvalue'
+                           )
+                        }}
+                        placeholder={`Enter ${columnName} prepend`}
+                     />
+                  </Form.Group>
+                  {concatType == 'concatData' && (
+                     <HelperText
+                        type="hint"
+                        message={`Enter comma separated values, for example: ${example}`}
+                     />
+                  )}
+               </HorizontalTabPanel>
+               <HorizontalTabPanel>
+                  <Flex
+                     container
+                     justifyContent="center"
+                     flexDirection="column"
+                  >
+                     <Flex container alignItems="center">
+                        <Text as="text1">Set {title} Null </Text>
+                        <Spacer xAxis size="10px" />
+                        <TextButton
+                           type="ghost"
+                           size="sm"
+                           onClick={() => {
+                              setBulkActions(prevState => {
+                                 delete prevState[columnName]
+                                 return prevState
+                              })
+                              setInitialBulkAction(prevState => ({
+                                 ...prevState,
+                                 [columnName]: '',
+                              }))
+                           }}
+                        >
+                           Clear
+                        </TextButton>
+                     </Flex>
+                     <Spacer size="10px" />
+                     <ButtonGroup align="left">
+                        <TextButton
+                           type="solid"
+                           size="sm"
+                           disabled={initialBulkAction[columnName] == null}
+                           onClick={() => {
+                              setInitialBulkAction(prevState => ({
+                                 ...prevState,
+                                 [columnName]: null,
+                              }))
+                              setBulkActions(prevState => ({
+                                 ...prevState,
+                                 [columnName]: null,
+                              }))
+                           }}
+                        >
+                           {initialBulkAction[columnName] == null
+                              ? 'Already'
+                              : 'Set'}{' '}
+                           Null
+                        </TextButton>
+                     </ButtonGroup>
+                  </Flex>
+               </HorizontalTabPanel>
+            </HorizontalTabPanels>
+         </HorizontalTabs>
+      </CollapsibleComponent>
    )
 }
 const CollapsibleComponent = ({ children, heading }) => (
